feat(logger): move to zap package
