+++ pkg/logger/module.go
@@ -1,8 +1,14 @@
 package logger
 
 import (
+	"github.com/alexfalkowski/go-service/pkg/logger/zap"
 	"go.uber.org/fx"
 )
 
-// Module for fx.
-var Module = fx.Provide(NewLogger)
+var (
+	// ZapLogger for fx.
+	ZapLogger = fx.Provide(zap.NewLogger)
+
+	// Module for fx.
+	Module = fx.Options(ZapLogger)
+)
