+++ pkg/logger/zap/logger.go
@@ -1,4 +1,4 @@
-package logger
+package zap
 
 import (
 	"context"
